+++ cx-portal/src/components/pages/EdcConnector/index.tsx
@@ -29,7 +29,6 @@ import {
   PageLoadingTable,
   PageSnackbar,
 } from 'cx-portal-shared-components'
-import SubHeaderTitle from 'components/shared/frame/SubHeaderTitle'
 import PictureWithText from 'components/shared/frame/PictureWithText'
 import AddConnectorOverlay from './AddConnectorOverlay'
 import { FormFieldsType } from 'components/pages/EdcConnector/AddConnectorOverlay'
@@ -85,6 +84,10 @@ const EdcConnector = () => {
   const [createDapsModalOpen, setCreateDapsModalOpen] = useState<boolean>(false)
   useState<boolean>(false)
   const [triggerDaps] = useTriggerDapsMutation()
+
+  const onStepChange = () => {
+    setAddConnectorOverlayCurrentStep(0)
+  }
 
   const onDelete = (row: ConnectorContentAPIResponse) => {
     setSelectedConnector(row)
@@ -256,15 +259,13 @@ const EdcConnector = () => {
         handleConfirmClick={onConfirmClick}
         onFormConfirmClick={onFormSubmit}
         loading={loading}
+        onStepChange={onStepChange}
       />
       <PageHeader
         title={t('content.edcconnector.headertitle')}
         topPage={false}
         headerHeight={200}
       />
-      <section>
-        <SubHeaderTitle title={'content.edcconnector.subheadertitle'} />
-      </section>
       <section className={'picture-with-text-section'}>
         <PictureWithText
           text={'content.edcconnector.imagetext'}