+++ CHANGELOG.md
@@ -48,6 +48,12 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
    * Filter feature enabled
    * View update in Details overlay
    * APIs added
+* Updates - Connector
+   * Upload title missing
+   * Change "Cancel" to "Back" Button
+   * Add "Cancel" icon in 2nd connector overlay
+   * Add tooltips
+   * Update titles
 
 
 ## 0.10.0